Fixed crash when using more than 255 leds

+++ WS2812ArtNet/esp8266.c
@@ -15,7 +15,7 @@ static inline uint32_t _getCycleCount(void) {
 }
 
 void ICACHE_RAM_ATTR espShow(
- uint8_t pin, uint8_t *pixels, uint32_t numBytes, boolean is800KHz) {
+ uint8_t pin, uint16_t *pixels, uint32_t numBytes, boolean is800KHz) {
 
 #define CYCLES_800_T0H  (F_CPU / 2500000) // 0.4us
 #define CYCLES_800_T1H  (F_CPU / 1250000) // 0.8us
